update pool user events
